Fix search for pkg-config on FreeBSD.
1) At least with FreeBSD's "pkg search", the search does *not* do a
prefix match, so if you look for "pkg-config", you can find packages
whose name is *not* pkg-config but that has "pkg-config" in the middle
of the name.  This means that we think we have a "pkg-config" package,
but we don't, and fail when we try to install it.

So we force a prefix match.

2) FreeBSD 11 doesn't have a "pkg-config" packate, but has a "pkgconf"
package.  If we don't find "pkg-config", look for "pkgconf".

Change-Id: Iad5ef9d5630981958830c03e4cb90fe2d01ce1d0
Reviewed-on: https://code.wireshark.org/review/29213
Reviewed-by: Guy Harris <guy@alum.mit.edu>