OAMPDU: fix dissection of DPOE TLV
When dissecting a DPOE TLV, dissector incorrectly attempts to
read a value when the value length is set to zero. This causes
the dissector to throw a malformed packet error. Added a check
for length zero and skip adding the value if length is zero.
Created an expert info output indicating that the TLV value is
zero length.

Change-Id: I313e20f6f436b9f0af3e6f82044964fe7c502485
Reviewed-on: https://code.wireshark.org/review/29270
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>