QUIC: implement packet coalescing (draft >= 11)
Messages with long headers have a payload length while short header
messages extend to the end of a datagram.

Create a new tvb for every message since the decryption routines assume
that the header starts from the beginning of the tvb.

quic_get_message_tvb tries to avoid throwing exceptions in most cases if
the (payload) length is too large such that dissection can proceed until
the payload is encountered. (One edge case is when the varint cannot
even be read, but then indeed the message is severely malformed.)

Change-Id: Id42f55dfd73905b709b218002b9ca660f6f703f1
Ping-Bug: 13881
Reviewed-on: https://code.wireshark.org/review/29607
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>