CMake: Fix build from tarball under certain conditions
RPM target has a hard dependency on a git checkout. When building
from a tarbal the cmake step fails on systems with rpmbuild and git
installed. Make that failure more rare by testing for the presence
of a .git subfolder, which 99% of the time implies a working git VCS.

Change-Id: I9e035fd0b192ae0e0fbe98599799f19dbf31263d
Reviewed-on: https://code.wireshark.org/review/29653
Petri-Dish: João Valverde <j@v6e.pt>
Tested-by: Petri Dish Buildbot
Reviewed-by: João Valverde <j@v6e.pt>