QUIC: remove draft -11 support
Draft -11 was more or less complete, but complicates PKN handling. Drop
it and update references to draft -13 (not the latest (-14) as we are
trying to get -12 and -13 ready before moving on).

Drop the "Short Packet Type" field and add a new field for three
reserved bits. The Third, Fourth and Google QUIC Demultiplexing bits are
not shown now for simplicity reasons.

Change-Id: Ibf4e281a54488aab14f5c9ca48ecc532e9701b12
Ping-Bug: 13881
Reviewed-on: https://code.wireshark.org/review/29668
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Reviewed-by: Peter Wu <peter@lekensteyn.nl>