Wiretap isn't meant to be a packet-capture library [any more].
Wiretap was originally being writen to be a packet-capture library which
might eventually replace libpcap but that's no longer the case: it's now
meant to be a library for accessing multiple types of files.

As such, don't describe it in the README as a packet-capture-library-to-be;
in fact since it's really internal to Wireshark don't mention it at all.

Don't discuss the fact that ancient (pre-2003) versions of zlib don't work
for Wireshark's purposes.

Change-Id: I2387d0ecc65a66c21dd7db3a38b8fee7d32add7f
Reviewed-on: https://code.wireshark.org/review/29786
Reviewed-by: Anders Broman <a.broman58@gmail.com>