Separate signed and unsigned decimal UAT fields.
Most of them are unsigned; do the appropriate fetching, checking, and
writing-to-UAT-file for them.  Have separate macros and routines for the
one signed one, which is the drbid in the LTE MAC dissector.

Use the Wireshark string-to-number routines; they do the appropriate
bounds checking, and make sure unsigned numbers don't start with a -.

Change-Id: I4f137aa31d631c5b5622b2c320574b8ab3333f31
Reviewed-on: https://code.wireshark.org/review/30288
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>