RTPS: Added Endpoint Discovery reassembly
Added reassembly of the Endpoint Discovery Data and added the
writer id to the fragment_add_check() function as the data to be
used to identify fragments belonging to the same packet.

Also changed the order of the modification of the INFO column for
Discovery fragments (first add status, then the [Reassembled] tag).

Change-Id: I1e830a07cecbfffe8a8fa9514e54d2e0aefb59c1
Reviewed-on: https://code.wireshark.org/review/30379
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>