Add a tap for the radio dissector that passes the pseudo-header.
That's what tap listeners that want radio information should use.  (If
they need additional radio information, they should expand the
pseudo-header structure and have *all* the libwiretap readers and
radio-metadata dissectors that fill in that structure fill it in - or
not mark it as present if there's nothing to fill it in with.)

Also, note that the "nav" field appears to be the duration field from
the frame header.

Change-Id: Ia09c69f6463371c880f2bddfc58760699b796049
Reviewed-on: https://code.wireshark.org/review/30414
Reviewed-by: Guy Harris <guy@alum.mit.edu>